# NTA/src/main/resources/apiSpec.yaml
openapi: 3.1.0
info:
  title: NTA Travel & Booking API
  version: 1.0.0
  description: 'API for managing Patients, Bookings, Support Persons, Service Providers, and Accommodations.'
  contact:
    name: Hamish Phillips
    email: hamish@phillipsmusictech.co.nz
servers:
  - url: 'http://localhost:8080/api'
paths:
...
  '/bookings/{nhi}':
    get:
      summary: List bookings for a patient
      parameters:
        - name: nhi
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bookings for patient
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '404':
          description: Patient Not Found
      operationId: getPatientBookings
      description: Retrieve Patients Bookings.
      tags:
        - patients
        - bookings
    post:
      summary: ''
      operationId: addPatientBooking
      responses:
        '201':
          description: Created New Booking
        '409':
          description: Booking already exists
      description: Add a New Booking for a Patient
      tags:
        - patients
        - bookings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
        description: ''
      parameters: []
    parameters:
      - schema:
          type: string
        name: nhi
        in: path
        required: true
...
  '/bookings/{nhi}/{id}':
    get:
      summary: Get booking details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking Not Found
      operationId: getBookingById
      description: Retrieve booking by ID.
      tags:
        - bookings
    put:
      summary: Update a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
      responses:
        '200':
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking Not Found
      operationId: updateBookingById
      description: Update Booking by Id
      tags:
        - bookings
    delete:
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Booking cancelled
        '404':
          description: Booking Not Found
      operationId: deleteBookingById
      description: Delete Booking by Id
      tags:
        - bookings
    parameters:
      - schema:
          type: string
        name: nhi
        in: path
        required: true
      - schema:
          type: string
        name: id
        in: path
        required: true
  '/bookings/{id}/notes':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Get Booking Notes
      tags:
        - bookings
        - notes
      responses:
        '204':
          description: No Content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
                  x-stoplight:
                    id: ke80m2f0zdq17
        '404':
          description: Booking Not Found
      operationId: getBookingNotes
      description: Retrieve the Booking Notes for a booking
    post:
      summary: Add Booking Note
      operationId: addBookingNote
      responses:
        '200':
          description: OK
          content: {}
        '404':
          description: Booking Not Found
      description: Add a Note to Booking
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      tags:
        - bookings
        - notes
...
components:
  schemas:
    Patient:
      type: object
      properties:
        nhi:
          type: string
          maxLength: 7
        ntaNumber:
          type: string
          maxLength: 7
        firstName:
          type: string
          maxLength: 30
        surname:
          type: string
          maxLength: 30
        notes:
          type: array
          x-stoplight:
            id: a5b9uo2h5tn2b
          items:
            $ref: '#/components/schemas/Note'
            x-stoplight:
              id: 7118i73waf800
        supportProviders:
          x-stoplight:
            id: tu0kke7d8sy77
          type: array
          items:
            $ref: '#/components/schemas/PatientServiceProvider'
            x-stoplight:
              id: 0i20xl1qen82s
      required:
        - nhi
        - firstName
        - surname
    SupportPerson:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
          maxLength: 30
        surname:
          type: string
          maxLength: 30
        coveredByNta:
          type: boolean
        patient:
          $ref: '#/components/schemas/Patient'
          x-stoplight:
            id: sv805lhoqkij2
      required:
        - id
        - firstName
        - surname
        - coveredByNta
    Booking:
      type: object
      properties:
        id:
          type: integer
        dateOfDeparture:
          type: string
          format: date
        dateOfReturn:
          type: string
          format: date
          nullable: true
        destination:
          type: string
          x-stoplight:
            id: 9ralr1tinlcdr
        bookingStatus:
          type: string
          maxLength: 15
        estimatedCost:
          type: number
          format: float
        estimatedCostForPatient:
          type: number
          x-stoplight:
            id: eu7asntsiivnn
          format: float
        bookingCreatedAt:
          type: string
          format: date-time
        serviceProvider:
          $ref: '#/components/schemas/ServiceProvider'
          x-stoplight:
            id: zwt73mcuom7en
        accommodation:
          $ref: '#/components/schemas/Accommodation'
          x-stoplight:
            id: omsd7xtb6by5i
        supportPersons:
          type: array
          items:
            $ref: '#/components/schemas/BookedSupportPerson'
        notes:
          x-stoplight:
            id: q1n5a1szwt3d5
          type: array
          items:
            $ref: '#/components/schemas/Note'
            x-stoplight:
              id: 0wfqry889bsg8
        patientNhi:
          $ref: '#/components/schemas/Patient'
          x-stoplight:
            id: cf6695mumxqcj
      required:
        - dateOfDeparture
        - bookingStatus
        - estimatedCost
        - bookingCreatedAt
        - serviceProvider
        - patientNhi
    ServiceProvider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          maxLength: 50
        address:
          type: string
          maxLength: 100
      required:
        - name
        - address
    Accommodation:
      type: object
      properties:
        address:
          type: string
          maxLength: 100
        name:
          type: string
          maxLength: 50
      required:
        - address
        - name
    BookedSupportPerson:
      title: BookedSupportPerson
      x-stoplight:
        id: wjwp2ksall9pj
      type: object
      properties:
        supportPerson:
          $ref: '#/components/schemas/SupportPerson'
          x-stoplight:
            id: w5a3zm38w0tjk
    Note:
      title: Note
      x-stoplight:
        id: t1hhkqpm11vef
      type: object
      description: Notes Model
      properties:
        id:
          type: string
        message:
          type: string
          x-stoplight:
            id: bm45upjpx1c6g
        booking:
          $ref: '#/components/schemas/Booking'
          x-stoplight:
            id: k5x237crt9duh
        patient:
          $ref: '#/components/schemas/Patient'
          x-stoplight:
            id: 3qjg5h0ofe2ju
      required:
        - message
    PatientServiceProvider:
      title: PatientServiceProvider
      x-stoplight:
        id: ya45sz5s3ky2n
      type: object
      properties:
        id:
          type: string
          x-stoplight:
            id: 8tjjqoynmkgdq
        serviceProvider:
          $ref: '#/components/schemas/ServiceProvider'
          x-stoplight:
            id: gaxqo9yozbgo0
        frequency:
          type: string
          x-stoplight:
            id: pqjt85fm823dg
      required:
        - serviceProvider
        - frequency
...